clk: multi-path parent rate propagation
The coordinated clock rates feature will allow multiple clks to have
their rates changed at once. A prerequisite to this feature is the
ability to have sibling clocks changed as a result of calling
clk_set_rate. Such clks might not even share a common parent, but must
be coordinated.

This patch expands clk_calc_new_rates from a single list walk up the
parent chain to a series of branching list walks as required by affected
sibling clks.

This is primarily achieved by keeping a list of the top-most clks found
in the tree traversal, as opposed to the old way of tracking a single
top-most clk.

Signed-off-by: Michael Turquette <mturquette@baylibre.com>