clk: mediatek: add CPU DVFS coordinated clock rate domain
Some Mediatek SoCs have special programming flow to be followed to make
CPU DVFS work. For CPU DVFS domain of these SoCs, two voltage inputs,
Vproc and Vsram are supplied by two regulators. In those cases that the
two regulators come from different PMICs, when scaling voltage inputs of
the CPU DVFS domain, the voltages of two regulator inputs need to be
controlled by software explicitly under the a limitation:

        100mV < Vsram - Vproc < 200mV

which is called 'voltage tracking' mechanism. And when scaling the
frequency of DVFS domain clock, the input MUX need to be parented to
another "intermediate" stable PLL first and reparented to the original
PLL once the original PLL is stable at the target frequency.

This patch uses the new "coordinated clock rate change" feature of
common clock domain and implements the CPU DVFS coordinated clock
domain which includes a PLL clock supplys the input clock of CPU DVFS
domain and a MUX clock to switch between the oringal PLL clock and the
intermediate PLL clock during PLL rate changing. The voltage tracking
mechanism is also included in this patch.

Signed-off-by: Pi-Cheng Chen <pi-cheng.chen@linaro.org>